fix: data format change in pirate weather

+++ src/main/java/com/bigboxer23/solar_moon/weather/PirateWeatherDataResponse.java
@@ -8,7 +8,7 @@ public class PirateWeatherDataResponse {
 	private double latitude;
 	private double longitude;
 	private String timezone;
-	private int offset;
-	private int elevation;
+	private double offset;
+	private double elevation;
 	private PirateWeatherData currently;
 }
